+++ packages/shared/src/index.ts
@@ -20,6 +20,7 @@ export * from './registrar.js'
 export * from './remoteProxy.js'
 export * from './scheduler.js'
 export * from './setup.js'
+export * from './snowbridge/governance.js'
 export * from './staking.js'
 export * from './system.js'
 export * from './treasury.js'
